pyproject.toml
[tool.poetry]
name = "parking"
version = "0.1.0"
description = ""
authors = ["guilherme <guilherme-lopes58@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
django = "^5.2"
djangorestframework = "^3.16.0"
django-jazzmin = "^3.0.1"
djangorestframework-simplejwt = "^5.5.0"
django-rql = "^4.4.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
flake8-pyproject = "^1.2.3"

[tool.flake8]
exclude = ['venv', '.venv', '__pycache__', '__init__.py']
ignore = 'E501'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
